Change ingredient quantity to double (plus fix controller)

+++ DAOs/Mappers/IngredientMapper.cs
@@ -20,7 +20,7 @@ public class IngredientMapper : AbstractMapper<Ingredient>
 
         return new Ingredient(
             reader.GetString(reader.GetOrdinal("name")),
-            reader.GetInt32(reader.GetOrdinal("quantity")),
+            reader.GetDouble(reader.GetOrdinal("quantity")),
             uom,
             notes
         );
